1.77 (#20)
* fix for broken XML reading
upped versions

* a more solid writeDatabase routine

* resync after corrupt database xml will now NOT re-download files if they still exist on the filesystem, and have matching content hashes.

* resync fixes

* can now cancel downloads, and can start beatstart without pedal

* manual beatstart starting

* manual beatstart starting

* now ignores dead google drive shortcut links

* no longer replaces accidentals in unrecognized chords

* no longer replaces accidentals in unrecognized chords

* added ellipsis/trailing-punctuation prefs

* don't use substitution in midi/setlist file parsing

* added sortable tags for year and icon (still to implement icon display)

* implemented song icons

* implemented song icons

* song parser ignores year and icon tags (only used in song list)

* song parser ignores year and icon tags (only used in song list)

* swapped year and key order in display

* code tidy

* added customizable song icon display position

* added missing icon

* code tidy

* added padding between left song icon and title

* test fixes

* code tidy

* gradle up

* gradle up

* dropbox up

* changed "show song title" to a more flexible list of possible things to show over the beat counter

* moved ActivityResultLauncher registration out of onCreate

* upped versionCode

* can now launch Google Docs editor directly from long-click menu

* can now launch Google Docs editor directly from long-click menu

* text overlay fix

* code tidy and version up

* code tidy

* code tidy

* code tidy

* code tidy

* icon update

* Create README.md

* start of docs import

* resized

* more docs

* added demo files from demo file repository

* added icons instead of links for donation support

* anchor test

* anchors

* readme updates

* readme updates

* update

---------

+++ app/src/main/kotlin/com/stevenfrew/beatprompter/storage/dropbox/DropboxStorage.kt
@@ -24,6 +24,8 @@ import com.stevenfrew.beatprompter.storage.StorageListener
 import com.stevenfrew.beatprompter.storage.StorageType
 import com.stevenfrew.beatprompter.storage.SuccessfulDownloadResult
 import com.stevenfrew.beatprompter.util.Utils
+import com.stevenfrew.beatprompter.util.getHash
+import com.stevenfrew.beatprompter.util.toHashString
 import io.reactivex.subjects.PublishSubject
 import org.apache.commons.io.FilenameUtils
 import java.io.File
@@ -34,6 +36,8 @@ import java.io.FileOutputStream
  */
 class DropboxStorage(parentFragment: Fragment) :
 	Storage(parentFragment, StorageType.Dropbox) {
+
+	private val hashBuffer = ByteArray(HASH_BUFFER_SIZE)
 
 	private val requestConfig = DbxRequestConfig.newBuilder(BeatPrompter.APP_NAME)
 		.build()
@@ -77,6 +81,7 @@ class DropboxStorage(parentFragment: Fragment) :
 					val localFile = downloadDropboxFile(client, metadata, targetFile)
 					val updatedCloudFile = FileInfo(
 						it.id, metadata.name, metadata.serverModified,
+						metadata.contentHash ?: "",
 						it.subfolderIds
 					)
 					SuccessfulDownloadResult(updatedCloudFile, localFile)
@@ -92,13 +97,39 @@ class DropboxStorage(parentFragment: Fragment) :
 		}
 	}
 
+	fun ByteArray.getSha256Hash() = getHash("SHA-256")
+
+	fun File.getDropboxHash(): String {
+		val inputStream = this.inputStream()
+		val sha256s = mutableListOf<ByteArray>()
+		while (true) {
+			val bytesRead = inputStream.read(hashBuffer)
+			if (bytesRead > 0) {
+				val bytes = if (bytesRead == HASH_BUFFER_SIZE) hashBuffer else hashBuffer.copyOfRange(
+					0,
+					bytesRead
+				)
+				sha256s.add(bytes.getSha256Hash())
+			}
+			if (bytesRead != HASH_BUFFER_SIZE)
+				break
+		}
+		val allBytes = sha256s.reduce { a1, a2 -> a1 + a2 }
+		val finalHash = allBytes.getSha256Hash()
+		val finalHashString = finalHash.toHashString(64)
+		return finalHashString
+	}
+
 	private fun downloadDropboxFile(client: DbxClientV2, file: FileMetadata, localFile: File): File =
 		localFile.also {
-			FileOutputStream(it).use { stream ->
-				client.files().download(file.id).use { downloader ->
-					downloader.download(stream)
+			// If we already have the file, we must be downloading cos the database
+			// was corrupted. The existing file might be valid.
+			if (!localFile.exists() || localFile.getDropboxHash() != file.contentHash)
+				FileOutputStream(it).use { stream ->
+					client.files().download(file.id).use { downloader ->
+						downloader.download(stream)
+					}
 				}
-			}
 		}
 
 	private fun readFolderContents(
@@ -141,6 +172,7 @@ class DropboxStorage(parentFragment: Fragment) :
 								itemSource.onNext(
 									FileInfo(
 										metadata.id, metadata.name, metadata.serverModified,
+										metadata.contentHash ?: "",
 										if (folderToSearch.parentFolder == null) "" else currentFolderID
 									)
 								)
@@ -148,9 +180,9 @@ class DropboxStorage(parentFragment: Fragment) :
 							Logger.log("Adding folder to list of folders to query ...")
 							val newFolder = FolderInfo(
 								folderToSearch,
-								metadata.getPathLower(),
+								metadata.getPathLower() ?: "",
 								metadata.getName(),
-								metadata.getPathDisplay()
+								metadata.getPathDisplay() ?: ""
 							)
 							if (recurseSubfolders)
 								foldersToSearch.add(newFolder)
@@ -202,9 +234,9 @@ class DropboxStorage(parentFragment: Fragment) :
 						DROPBOX_APP_KEY
 					)
 					updateDropboxCredentials(newCred)
-				} catch (authEx: DbxOAuthException) {
+				} catch (_: DbxOAuthException) {
 					return null
-				} catch (ex: DbxException) {
+				} catch (_: DbxException) {
 					return null
 				}
 			} else cred
@@ -263,6 +295,7 @@ class DropboxStorage(parentFragment: Fragment) :
 
 	companion object {
 		const val DROPBOX_CACHE_FOLDER_NAME = "dropbox"
+		private const val HASH_BUFFER_SIZE = 4 * 1024 * 1024
 
 		@Suppress("SpellCheckingInspection")
 		private const val DROPBOX_APP_KEY = "hay1puzmg41f02r"
